update chat separator
