+++ src/lib/components/chat/Chat.svelte
@@ -9,7 +9,7 @@
 	import Event from "../message/Event.svelte";
 	import Notification from "../message/Notification.svelte";
 	import UserMessage from "../message/UserMessage.svelte";
-	import Separator from "./Separator.svelte";
+	import ChatSeparator from "./ChatSeparator.svelte";
 
 	interface Props {
 		class?: string;
@@ -31,12 +31,7 @@
 		if (!scrollingPaused) scrollToEnd();
 	});
 
-	const newMessageCount = $derived.by(() => {
-		if (!list) return 0;
-
-		return chat.messages.length - countSnapshot;
-	});
-
+	const newMessageCount = $derived(chat.messages.length - countSnapshot);
 	const hasNew = $derived(newMessageCount > 0);
 
 	$effect(() => {
@@ -85,9 +80,11 @@
 	{#if scrollingPaused}
 		<div class="absolute inset-x-0 bottom-4 z-10 flex justify-center">
 			<button
-				class="flex items-center rounded-full border bg-twitch/40 p-1.5 text-xs font-medium shadow-sm backdrop-blur-lg transition-[padding] duration-200 ease-out hover:bg-twitch/60 data-[new=true]:pr-3"
+				class={[
+					"flex items-center rounded-full border bg-twitch/40 p-1.5 text-xs font-medium shadow-sm backdrop-blur-lg transition-[background-color,padding] duration-200 ease-out hover:bg-twitch/60",
+					hasNew && "pr-3",
+				]}
 				type="button"
-				data-new={hasNew}
 				onclick={scrollToEnd}
 				transition:fly={{ y: 16, duration: 200 }}
 			>
@@ -118,16 +115,13 @@
 				{@const isNewDay = prev && prev.timestamp.getDate() !== message.timestamp.getDate()}
 
 				{#if isNewDay}
-					<Separator>
-						<time
-							class="text-muted-foreground/90"
-							datetime={message.timestamp.toISOString()}
-						>
+					<ChatSeparator>
+						<time datetime={message.timestamp.toISOString()}>
 							{message.timestamp.toLocaleDateString(navigator.languages, {
 								dateStyle: "long",
 							})}
 						</time>
-					</Separator>
+					</ChatSeparator>
 				{/if}
 
 				{#if message.isEvent()}
@@ -146,11 +140,11 @@
 				{@const nextRecent = next && (next.isEvent() || next.isUser()) && next.recent}
 
 				{#if message === lastRead && next && settings.state["chat.newSeparator"]}
-					<Separator class="text-red-400">New messages</Separator>
+					<ChatSeparator class="text-twitch-400">New messages</ChatSeparator>
 				{/if}
 
 				{#if message.recent && !nextRecent && settings.state["chat.messages.history.separator"]}
-					<Separator class="text-red-400">Live messages</Separator>
+					<ChatSeparator class="text-red-400">Live messages</ChatSeparator>
 				{/if}
 			{:else if message.isComponent()}
 				<message.component {...message.props} />
